cleanup and escape notification bodies
